Apply review feedback: BIDS-valid scenario names, centralized fixtures, copy_state_file rename
- Give every example entry a BIDS-valid dandi_path (single dash per entity): two-word
  scenarios become sub-<word>/ses-<word> (e.g. sub-failed/ses-repeated,
  sub-two/ses-attempts), and the path-mismatch entry uses sourcedata/sub-test+bids.
- Move the reusable test-local fixtures (processing_directory, populated_queue,
  _dandi_api_key) into conftest.py.
- Rename the install_state_file fixture to copy_state_file, reflecting that the queue
  functions only read state.jsonl (never mutate it); the copy exists so the
  directory-based API finds the file in an isolated tmp queue.
