+++ tests/dandi_compute_code/queue/example_state_files/README.md
@@ -24,10 +24,10 @@ The attempt lifecycle is encoded by the presence flags:
 | `sub-pending` | Prepared but never submitted. |
 | `sub-running` | Logs present, no output yet. |
 | `sub-successful` | Output present, with a known source-asset size (120 bytes). |
-| `sub-failed-repeated` (attempts 1 and 2) | Two failed attempts of one asset, reaching `max_fail_per_dandiset` for Dandiset 000001 (mapped to `asset-aaa`). |
+| `sub-failed/ses-repeated` (attempts 1 and 2) | Two failed attempts of one asset, reaching `max_fail_per_dandiset` for Dandiset 000001 (mapped to `asset-aaa`). |
 | `sub-fresh` (Dandiset 000002) | A queued asset in another Dandiset with no failures (mapped to `asset-bbb`). |
-| `sub-with-session/ses-recording` | A queued attempt nested under a session directory. |
-| `sub-sole-attempt/ses-recording` (Dandiset 001371) | The sole attempt in its pipeline/version tree, so empty parents are pruned on removal. Also covers the legacy nested layout. |
-| `sub-two-attempts` (attempts 1 and 2) | Attempt 1 queued, attempt 2 completed, so the shared parent is kept after the queued attempt is removed. |
-| `sub-already-submitted` | Queued in state, but a submitted marker exists on disk, so it is left alone. |
-| `sourcedata` (Dandiset 001849) | A queued attempt whose recorded `dandi_path` differs from the on-disk attempt layout, exercising fallback attempt-directory resolution. |
+| `sub-session/ses-recording` | A queued attempt nested under a session directory. |
+| `sub-sole/ses-attempt` (Dandiset 001371) | The sole attempt in its pipeline/version tree, so empty parents are pruned on removal. Also covers the legacy nested layout. |
+| `sub-two/ses-attempts` (attempts 1 and 2) | Attempt 1 queued, attempt 2 completed, so the shared parent is kept after the queued attempt is removed. |
+| `sub-already/ses-submitted` | Queued in state, but a submitted marker exists on disk, so it is left alone. |
+| `sourcedata/sub-test+bids` (Dandiset 001849) | A queued attempt whose recorded `dandi_path` differs from the on-disk attempt layout, exercising fallback attempt-directory resolution. |
