+++ tests/dandi_compute_code/queue/test_clean_unsubmitted_capsules.py
@@ -1,18 +1,11 @@
 import os
 import pathlib
-from collections.abc import Callable, Iterator
+from collections.abc import Callable
 from unittest import mock
 
 import pytest
 
 from dandi_compute_code.queue import JobEntry, clean_unsubmitted_capsules
-
-
-@pytest.fixture
-def _dandi_api_key() -> Iterator[None]:
-    """clean_unsubmitted_capsules requires DANDI_API_KEY; provide a dummy value."""
-    with mock.patch.dict(os.environ, {"DANDI_API_KEY": "test-key"}):
-        yield
 
 
 @pytest.mark.ai_generated
@@ -26,7 +19,7 @@ def test_clean_unsubmitted_capsules_raises_without_dandi_api_key(tmp_path: pathl
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_removes_queued_directories(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -34,7 +27,7 @@ def test_clean_unsubmitted_capsules_removes_queued_directories(
 ) -> None:
     """clean_unsubmitted_capsules removes capsule dirs that are queued (code, no logs, no output)."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-pending"))
 
     with mock.patch("subprocess.run") as mock_run:
@@ -48,7 +41,7 @@ def test_clean_unsubmitted_capsules_removes_queued_directories(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_skips_entries_with_output(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -56,7 +49,7 @@ def test_clean_unsubmitted_capsules_skips_entries_with_output(
 ) -> None:
     """clean_unsubmitted_capsules does not remove capsules that already have output."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     completed_dir = create_attempt_directory(
         base_dir=dandiset_dir, entry=entry_for("sub-successful"), with_logs=True, with_output=True
     )
@@ -72,7 +65,7 @@ def test_clean_unsubmitted_capsules_skips_entries_with_output(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_skips_entries_with_logs(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -80,7 +73,7 @@ def test_clean_unsubmitted_capsules_skips_entries_with_logs(
 ) -> None:
     """clean_unsubmitted_capsules does not remove capsules that have logs (already run)."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     failed_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-running"), with_logs=True)
 
     with mock.patch("subprocess.run") as mock_run:
@@ -94,7 +87,7 @@ def test_clean_unsubmitted_capsules_skips_entries_with_logs(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_ignores_dataset_description_in_logs(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -102,7 +95,7 @@ def test_clean_unsubmitted_capsules_ignores_dataset_description_in_logs(
 ) -> None:
     """A logs/ directory holding only dataset_description.json does not protect a queued capsule."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-pending"))
     logs_dir = queued_dir / "logs"
     logs_dir.mkdir()
@@ -118,7 +111,7 @@ def test_clean_unsubmitted_capsules_ignores_dataset_description_in_logs(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_skips_entries_with_submitted_marker(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -126,7 +119,7 @@ def test_clean_unsubmitted_capsules_skips_entries_with_submitted_marker(
 ) -> None:
     """clean_unsubmitted_capsules does not remove capsules with a submitted marker file."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-pending"), submitted=True)
 
     with mock.patch("subprocess.run") as mock_run:
@@ -140,13 +133,13 @@ def test_clean_unsubmitted_capsules_skips_entries_with_submitted_marker(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_returns_empty_list_when_nothing_queued(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
     _dandi_api_key: None,
 ) -> None:
     """clean_unsubmitted_capsules returns an empty list when nothing is materialized on disk."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
 
     removed = clean_unsubmitted_capsules(dandiset_directory=dandiset_dir, queue_directory=queue_directory)
 
@@ -156,7 +149,7 @@ def test_clean_unsubmitted_capsules_returns_empty_list_when_nothing_queued(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_handles_session_in_path(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -164,8 +157,8 @@ def test_clean_unsubmitted_capsules_handles_session_in_path(
 ) -> None:
     """clean_unsubmitted_capsules correctly handles attempt dirs with a session component."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
-    queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-with-session/ses-recording"))
+    copy_state_file(queue_directory=queue_directory)
+    queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-session/ses-recording"))
 
     with mock.patch("subprocess.run"):
         removed = clean_unsubmitted_capsules(dandiset_directory=dandiset_dir, queue_directory=queue_directory)
@@ -177,7 +170,7 @@ def test_clean_unsubmitted_capsules_handles_session_in_path(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_removes_empty_parent_directories(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -185,8 +178,8 @@ def test_clean_unsubmitted_capsules_removes_empty_parent_directories(
 ) -> None:
     """clean_unsubmitted_capsules removes empty pipeline/session dirs after last capsule removal."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
-    queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-sole-attempt/ses-recording"))
+    copy_state_file(queue_directory=queue_directory)
+    queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-sole/ses-attempt"))
     pipeline_dir = queued_dir.parent
     session_dir = pipeline_dir.parent
 
@@ -201,7 +194,7 @@ def test_clean_unsubmitted_capsules_removes_empty_parent_directories(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_keeps_non_empty_parent_directories(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -209,10 +202,10 @@ def test_clean_unsubmitted_capsules_keeps_non_empty_parent_directories(
 ) -> None:
     """clean_unsubmitted_capsules keeps pipeline/version dirs when a sibling attempt remains."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
-    queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-two-attempts", attempt=1))
+    copy_state_file(queue_directory=queue_directory)
+    queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-two/ses-attempts", attempt=1))
     remaining_dir = create_attempt_directory(
-        base_dir=dandiset_dir, entry=entry_for("sub-two-attempts", attempt=2), with_output=True
+        base_dir=dandiset_dir, entry=entry_for("sub-two/ses-attempts", attempt=2), with_output=True
     )
     pipeline_dir = queued_dir.parent
 
@@ -227,7 +220,7 @@ def test_clean_unsubmitted_capsules_keeps_non_empty_parent_directories(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_removes_legacy_nested_layout(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -235,9 +228,9 @@ def test_clean_unsubmitted_capsules_removes_legacy_nested_layout(
 ) -> None:
     """clean_unsubmitted_capsules removes queued capsules in the legacy nested layout."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     queued_dir = create_attempt_directory(
-        base_dir=dandiset_dir, entry=entry_for("sub-sole-attempt/ses-recording"), legacy_nested=True
+        base_dir=dandiset_dir, entry=entry_for("sub-sole/ses-attempt"), legacy_nested=True
     )
     version_dir = queued_dir.parent
     pipeline_dir = version_dir.parent
@@ -254,7 +247,7 @@ def test_clean_unsubmitted_capsules_removes_legacy_nested_layout(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_removes_only_queued_not_submitted(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     entry_for: Callable[..., JobEntry],
     create_attempt_directory: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
@@ -262,10 +255,10 @@ def test_clean_unsubmitted_capsules_removes_only_queued_not_submitted(
 ) -> None:
     """clean_unsubmitted_capsules only removes queued capsules, leaving submitted ones intact."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
     queued_dir = create_attempt_directory(base_dir=dandiset_dir, entry=entry_for("sub-pending"))
     submitted_dir = create_attempt_directory(
-        base_dir=dandiset_dir, entry=entry_for("sub-already-submitted"), submitted=True
+        base_dir=dandiset_dir, entry=entry_for("sub-already/ses-submitted"), submitted=True
     )
 
     with mock.patch("subprocess.run"):
@@ -279,13 +272,13 @@ def test_clean_unsubmitted_capsules_removes_only_queued_not_submitted(
 @pytest.mark.ai_generated
 def test_clean_unsubmitted_capsules_removed_entry_via_fallback_attempt_resolution(
     queue_directory: pathlib.Path,
-    install_state_file: Callable[..., pathlib.Path],
+    copy_state_file: Callable[..., pathlib.Path],
     tmp_path: pathlib.Path,
     _dandi_api_key: None,
 ) -> None:
     """clean_unsubmitted_capsules removes queued entry when dandi_path differs from on-disk attempt path."""
     dandiset_dir = tmp_path / "dandiset"
-    install_state_file(queue_directory=queue_directory)
+    copy_state_file(queue_directory=queue_directory)
 
     # The "sourcedata" entry's on-disk attempt lives under sub-mouse01, so it must be
     # located via fallback resolution rather than the recorded dandi_path.
